infrastructure/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
aws-cdk.aws-ec2>=2.0.0
aws-cdk.aws-rds>=2.0.0
aws-cdk.aws-secretsmanager>=2.0.0 